Reached stable runtime
The functor and hierarchy classes need to be refactored for perfect-forwarding of
constness and reference types. also it would be great if we could use as much constexpr as possible
and be able to use as many const objects as possible. Widgets should be entirely static objects,
they are initialized once, put their data in vectors and never change any of their data again,
but use other data to operate on their data stored in the vectors.