Cleaned up, formatting
Removed Element hierarchy functionality from Quad class,
moved widget declarations into widget.h
Planning to implement a single Element wrapper to put the base elements into.
It would be useful to have some funtionality internal to an Element hierarchy
but it cant be expected that all this functionality is availible in all types
used as Base Elements. Ideally, there should be a uniform interface for each element
in a hierarchy, no matter if it is a group of elements or just a single element.