Cleared up errors in line module
line vertices are now a pair of indices (pos & color)
lines are a pair of indices to two vertices

the line buffer is used as the element buffer for the shader
which uses it to access the vertex array. this causes a lot of random access
in is not needed in most cases (lines using two dots) although a little bit of
memory is saved... never mind, actually a line takes up just as much memory as
an extra vertex.