Began functions for loading ttf files into bitmap atlas and glyphs
Initial commit for the font loader
The font loader is a module which uses freetype to load character bitmaps and metrics from ttf files.
The bitmap for each character will be stored in an atlas for faster rendering and the metrics will be used
for formatting the characters for well readable text.
The text will mainly be used for the GUI, including text journal entries, so it should be rather simplistic
but well readable. (think terminals)