Integrated Font class into master source
Font is a class which handles reading and writing font faces from/to files.
It stores glyph bitmaps in an atlas, glyph metrics and (not yet) kerning values.
It can read those values through FreeType and write the data in bare bit format into
a .font file (and read that)