Expanded Itr class
The class is still not coming up to its purpose, since references cannot be reassigned.
It would be nice to embed the reference to the target container itself as a template parameter,
so that it is static for all objects of a template specific instantiation.
This would be just like declaring an ID wrapper for each conceptual container but without the
code duplication.