Refactor Event handling
Removed distinction between Events and EventSlots

Event is a class template which holds a signature of its type and an ID
of a Signal which it triggers when it event occurs. pushEvent() pushes
an event into the event queue of an Event class and that class can then
compare all of its instance's signatures with that queue. When an Event
signature matches an event in the queue, the Event triggers its signal.