Rename Listener to EventListener and add SignalListener prototype
EventListeners listen directly to events pushed by pushEvent and set
their state when they occurred.
SignalListeners should be able to listen to any number of
Event/SignalListeners and they will use a logical condition to set their
own state based on the states of the Listeners they are listening to.