Add signals to Widgets and refactor
Widgets can now give signal handles for enter and leave events.
Widgets now also store color in the QuadWidgets which are at the base of
the hierarchy of a widget.
These QuadWidgets are templates and can use different color types, this
allows for constant, dynamic colors, and textures. Or actually you could
also create a type of color that mixes textures and colors. basically
anything.