Make Key and Mouse Signals IDs, so that they can be fully const
because ButtonSignals uses IDs itself, ButtonSignals stored at
namespace-level can't be const, because they can't be initialized at
definition, because that happens in an undefined order, which means that
the ID containers may be initialized after the ButtonSignals have been
initialized, which would cause the data to be overwritten.

At least that is what I think so far... I am not sure though. Because
now that I think about it, the ID containers should be defined as soon
as they are used...