Remove WidgetLayout, expand Widget::Preset
Widgets now take more parameters in their constructor, thus having a
simplet template interface because more customized processing can be
kept in the widget configuration (the wrapper for Widget)

Widgets can now not be used alone anymore, but they need to be
initalized with a quad for each of their elements.

Also colors are not stored seperately anymore, but are simply the
Presets for QuadElements.

Move- and ResizePolicies are also passed to Widget.

Designers of a Widget can expand the Widget configuration however
they like, but can never change the constructor or how the base widget
is initialized.

The core initialization process is now:
- Define Preset with MovePolicy, ResizePolicy and SubPresets
- Initialize Wrapper with a quad and the Preset
- Wrapper generates Quads based using a member function
- Base Widget is initalized with original quad, the Quads and the Preset
- The Widget copies the Move- and ResizePolicy from the Preset
- The Widget uses the Quads and the SubPresets to initialize each of its
  Elements
- Recurse to step 2 if the Element is not a QuadElement
- Otherwise, QuadElements use the Quad and the Preset, which is a color,
  to create a colored Quad on the screen