Add ManagedID and use it in QuadElements
ManagedID is a move-only type storing a managed index into a vector of a
type. "Managed" means that it will always remain valid, even if elements
in front of it are deleted from the vector.

For this, WidgetSignals had to be enabled for now, since they dont
support move only types yet.