Use basic IDs in QuadElements again
ManagedIDs cannot be copied (and should not), so this would make the
entire Widget not copyable. This makes it hard to use it with signals
and functors, for example because events have to store a QuadID to know
which quad to listen to, and functors may want to store a QuadID to know
which quad to color or move or whatever.