Add utils::Container and add Management Policies for utils::ID
utils::Container is a wrapper around std::vector and enables you to use
utils::IDs with different Containers. This way, rather than always
creating a new type for each individual use, an ID can be created for a
specific Container.

Also this commit adds Management policies to IDs and Containers. This
allows developers to create an ID type using a custom access or
destructing pattern and can be used to implement reference counted IDs
for example.