Have Widget inherit from BoundingBoxID
this allows you to treat a Widget as a BoundingBoxID, thus providing a
Quad interface for a widget. This way you can get the position and size
of the entire widget.

To-Do: make changes to a BoundingBoxID take effect on the actual widget
elements. Currently, when changing the width of a widget through its
BoundingBox, resize does not get called and the widget does not actually
change size.