Assemble Presets from Layouts and Colors
Colors and Layouts are now seperate, to make configuration easier. You
can now use the same colors with a different layout for a Widget. The
Layout provides a function to create the subwidget quads at
initialization and that can use any member variables of the layout,
making it highly customizable.