Implement first version of dynamic Text
This makes it possible to append text to textboxe which will appear on
the screen. in this version all Textboxes get recalculated and reuploaded every
frame, which could possibly be optimized.
Notes:
there are a few different types of text and the implementation should be
able to optimize accordingly. some Textboxes always contain the same
text and may only be hidden entirely. other textboxes will have dynamic
text which can grow or shrink in size, other textboxes may also have
their content changed aswell, which will require all the following character's
positions to be translated, unless a monospaced font is used.
generally the text library should definitely add some functionality to move an
entire range of characters to move text.